Extract CurrentPlanDetailsView
